feat: add feature for showing fake error after clicking on trigger error button
