+++ class-components/src/index.css
@@ -72,8 +72,11 @@ p {
 
 .error-message {
   color: var(--accent-red);
-  font-weight: 400;
-  margin: 200px auto;
+  font-weight: 300;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
 }
 
 @media (hover: hover) {
